Safari-609.1.11.1/LayoutTests/crypto/subtle/aes-kw-generate-key-wrap-key-unwrap-key-expected.txt

Test wrapKey/unwrapKey with AES-KW using a generated key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Generating...
Wrapping...
PASS bytesToASCIIString(wrappedKey) is not "jnOw99oOZFLIEPMr"
Unwrapping...
PASS bytesToASCIIString(unwrappedKey) is rawKeyASCII
PASS successfullyParsed is true

TEST COMPLETE
